requirements.txt
pyecharts
pymysql
pyyaml
pandas